Hold the work on Docotor verification and changed the chat message from using signalR to send to HTTP

+++ doctor/verification/core-verification/src/main/java/com/seravian/core_verification/data/dto/request/GetVerificationsRequest.kt
@@ -1,5 +1,5 @@
 package com.seravian.core_verification.data.dto.request
 
 data class GetVerificationsRequest(
-    val requestId: Int
+    val requestId: Long
 )
